+++ DSCResources/MSFT_ScheduledTask/MSFT_ScheduledTask.psm1
@@ -649,7 +649,7 @@ function Set-TargetResource
             {
                 $triggerParameters.Add('AtLogOn', $true)
 
-                if (-not [System.String]::IsNullOrWhiteSpace($User))
+                if (-not [System.String]::IsNullOrWhiteSpace($User) -and $LogonType -ne 'Group')
                 {
                     $triggerParameters.Add('User', $User)
                 }
@@ -780,7 +780,6 @@ function Set-TargetResource
         elseif ($PSBoundParameters.ContainsKey('ExecuteAsCredential'))
         {
             $username = $ExecuteAsCredential.UserName
-            $registerArguments.Add('User', $username)
 
             # If the LogonType is not specified then set it to password
             if ([System.String]::IsNullOrEmpty($LogonType))
@@ -788,7 +787,12 @@ function Set-TargetResource
                 $LogonType = 'Password'
             }
 
-            if ($LogonType -notin ('Interactive', 'S4U'))
+            if ($LogonType -ne 'Group')
+            {
+                $registerArguments.Add('User', $username)
+            }
+
+            if ($LogonType -notin ('Interactive', 'S4U', 'Group'))
             {
                 # Only set the password if the LogonType is not interactive or S4U
                 $registerArguments.Add('Password', $ExecuteAsCredential.GetNetworkCredential().Password)
@@ -809,8 +813,16 @@ function Set-TargetResource
         # Prepare the principal arguments
         $principalArguments = @{
             Id        = 'Author'
-            UserId    = $username
-            LogonType = $LogonType
+        }
+
+        if ($LogonType -eq 'Group')
+        {
+            $principalArguments.GroupId = $username
+        }
+        else
+        {
+            $principalArguments.LogonType = $LogonType
+            $principalArguments.UserId = $username
         }
 
         # Set the Run Level if defined
@@ -1776,6 +1788,15 @@ function Get-CurrentResource
             $synchronizeAcrossTimeZone = $false
         }
 
+        if ($task.Principal.LogonType -ieq 'Group')
+        {
+            $PrincipalId = 'GroupId'
+        }
+        else
+        {
+            $PrincipalId = 'UserId'
+        }
+
         $result = @{
             TaskName                        = $task.TaskName
             TaskPath                        = $task.TaskPath
@@ -1788,7 +1809,7 @@ function Get-CurrentResource
             ActionWorkingPath               = $action.WorkingDirectory
             ScheduleType                    = $returnScheduleType
             RepeatInterval                  = ConvertTo-TimeSpanStringFromScheduledTaskString -TimeSpan $trigger.Repetition.Interval
-            ExecuteAsCredential             = $task.Principal.UserId
+            ExecuteAsCredential             = $task.Principal.$PrincipalId
             ExecuteAsGMSA                   = $task.Principal.UserId -replace '^.+\\|@.+', $null
             Enable                          = $settings.Enabled
             DaysInterval                    = [System.Uint32] $trigger.DaysInterval